chore: move functions to packages
